lgraph/pyproject.toml
[project]
name = "lgraph"
version = "0.1.0"
description = ""
authors = [
    {name = "sungminna",email = "sungmin.na330@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "langgraph (>=0.4.8,<0.5.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langgraph-supervisor (>=0.0.27,<0.0.28)",
    "langchain[openai] (>=0.3.26,<0.4.0)",
    "tavily-python (>=0.7.8,<0.8.0)",
    "fastapi (>=0.115.13,<0.116.0)",
    "uvicorn[standard] (>=0.34.3,<0.35.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "google-genai (>=1.22.0,<2.0.0)",
    "minio (>=7.2.15,<8.0.0)",
    "python-ffmpeg (>=2.0.12,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
